+++ Tests/AccessibilityTests/AccessibilityProject/Sources/TargetA/PublicActor.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+public actor PublicActor {
+  public init() { }
+  public func someFunc() {}
+}